requirements.txt
tqdm
fire
numpy
scipy
torch
datasets
pandas
bs4
gym
# hf
transformers
# llama
sentencepiece
git+https://github.com/facebookresearch/llama@llama_v1
fairscale
# openai model
openai
# game24
sympy
# visualization
requests~=2.31.0
PyYAML~=6.0
tarski~=0.8.2
setuptools~=67.8.0
#blocksworld
tarski
#quantization
peft~=0.5.0
optimum
ninja~=1.11.1
bitsandbytes~=0.41.1
azure-identity